backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.4.0
python-decouple==3.8
requests==2.31.0
